os-agent/pyproject.toml
[project]
name = "os-agent"
version = "1.0.0"
description = "操作系统智能代理 - 支持自然语言驱动的Linux服务器管理"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "OS Agent Team"}
]
dependencies = [
    # 核心框架 - LangChain 1.x
    "langchain-core>=1.0.0",
    "langchain-openai>=1.1.16",
    "langgraph>=1.0.0",
    "openai>=1.0.0",

    # Web服务
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "websockets>=12.0",

    # CLI
    "typer>=0.12.0",
    "rich>=13.0.0",

    # SSH远程连接
    "paramiko>=3.4.0",

    # 语音模块
    "faster-whisper>=1.0.0",
    "edge-tts>=6.1.0",
    "vosk>=0.3.45",

    # Web文件上传
    "python-multipart>=0.0.26",

    # 配置管理
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",

    # 日志
    "loguru>=0.7.0",

    # 工具
    "python-dotenv>=1.0.0",
    "aiofiles>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
os-agent = "src.interface.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
